html basic markup
marked up the html text. I'm least confident about how I marked up the thumbnail gallery. I considered a list but it seems like the images are just as important as the titles so I used figures. If I just let it be a text list and used background images, then a lot of content would be lost if the css doens't load for whatever reason.